Nimbus requires pow(Uint, uint64) but uint != uint64

+++ src/ttmath.nim
@@ -110,7 +110,7 @@ proc pow*(a: Int, b: int): Int =
   tmp.inplacePow(initInt[Int](b))
   result = tmp
 
-proc pow*(a: UInt, b: uint): UInt =
+proc pow*(a: UInt, b: uint64): UInt =
   var tmp = a
   tmp.inplacePow(initUInt[UInt](b))
   result = tmp
